Refactor: Profile 방문자 수 분리
