+++ src/main/java/supernova/whokie/profile/Profile.java
@@ -1,17 +1,7 @@
 package supernova.whokie.profile;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
-import jakarta.validation.constraints.NotNull;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import jakarta.persistence.*;
+import lombok.*;
 import supernova.whokie.user.Users;
 
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -28,12 +18,6 @@ public class Profile {
     @OneToOne
     @JoinColumn(name = "user_id")
     private Users users;
-
-    @NotNull
-    private Integer todayVisited;
-
-    @NotNull
-    private Integer totalVisited;
 
     private String description;
