+++ src/main/java/supernova/whokie/profile/controller/dto/ProfileResponse.java
@@ -7,8 +7,6 @@ public class ProfileResponse {
 
     @Builder
     public record Info(
-            int todayVisited,
-            int totalVisited,
             String description,
             String backgroundImageUrl,
             String name
@@ -16,8 +14,6 @@ public class ProfileResponse {
 
         public static ProfileResponse.Info from(ProfileModel.Info info) {
             return Info.builder()
-                    .todayVisited(info.todayVisited())
-                    .totalVisited(info.totalVisited())
                     .description(info.description())
                     .backgroundImageUrl(info.backgroundImageUrl())
                     .name(info.name())
