+++ src/main/java/supernova/whokie/profile/service/dto/ProfileModel.java
@@ -2,13 +2,12 @@ package supernova.whokie.profile.service.dto;
 
 import lombok.Builder;
 import supernova.whokie.profile.Profile;
+import supernova.whokie.profile.ProfileVisitCount;
 
 public class ProfileModel {
 
     @Builder
     public record Info(
-            int todayVisited,
-            int totalVisited,
             String description,
             String backgroundImageUrl,
             String name
@@ -16,11 +15,22 @@ public class ProfileModel {
 
         public static ProfileModel.Info from(Profile profile) {
             return Info.builder()
-                    .todayVisited(profile.getTodayVisited())
-                    .totalVisited(profile.getTotalVisited())
                     .description(profile.getDescription())
                     .backgroundImageUrl(profile.getBackgroundImageUrl())
                     .name(profile.getUsers().getName())
+                    .build();
+        }
+    }
+
+    @Builder
+    public record Visited(
+            int todayVisited,
+            int totalVisited
+    ) {
+        public static ProfileModel.Visited from(ProfileVisitCount visitCount) {
+            return Visited.builder()
+                    .todayVisited(visitCount.getDailyVisited())
+                    .totalVisited(visitCount.getTotalVisited())
                     .build();
         }
     }
