+++ src/test/java/supernova/whokie/profile/ProfileIntegrationTest.java
@@ -10,7 +10,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.web.servlet.MockMvc;
-import supernova.whokie.profile.infrastructure.ProfileRepository;
+import supernova.whokie.profile.infrastructure.repository.ProfileRepository;
 import supernova.whokie.user.Gender;
 import supernova.whokie.user.Role;
 import supernova.whokie.user.Users;
@@ -80,8 +80,6 @@ public class ProfileIntegrationTest {
     private Profile createProfile() {
         Profile profile = Profile.builder()
             .users(user)
-            .todayVisited(2)
-            .totalVisited(12)
             .description("test")
             .backgroundImageUrl("test")
             .build();
